+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
 
 **Improvements**:
 
+- Support custom filename in ``redmine.upload()``
 - Support for ``get()`` and ``update()`` operations for ``/my/account`` endpoint which doesn't require admin
   privileges by using ``me`` as an id, i.e. ``redmine.user.get('me')`` or ``redmine.user.update('me',firstname='John')``
   (requires Redmine >= 4.1.0)